Move top banner into own file
