+++ src/components/pages/educators/index.js
@@ -1,5 +1,6 @@
 import HowToJoin from './how-to-join';
 import ProgramBenefits from './program-benefits';
 import SignUpModal from './sign-up-modal';
+import TopBanner from './top-banner';
 
-export { HowToJoin, ProgramBenefits, SignUpModal };
+export { HowToJoin, ProgramBenefits, SignUpModal, TopBanner };
